backend/requirements.txt
xgboost==2.0.3
shap==0.44.1
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
faker==24.0.0
openai==1.12.0
python-dotenv==1.0.1
